Use Rust governance model as initial model
Sawtooth is adopting a governance model which is very similar to the process in
place for the Rust community. The Rust RFC that establishes their governance
model is at:

  https://github.com/rust-lang/rfcs/blob/master/text/1068-rust-governance.md

The Rust Governance RFC is copied from the above repository and copyright of
the text can be determined in that repository. The repository is dual-licensed,
and we are using the RFC under Apache 2.0.

The LICENSE-APACHE file from the above repository is included as LICENSE in
this repository.

Signed-off-by: Adam Ludvik <ludvik@bitwise.io>