Proposes Sawtooth Sabre transaction family
Proposes Sawtooth Sabre, a transaction family which implements on-chain
smart contracts executed in a WebAssembly virtual machine.

WebAssembly (Wasm) is a stack-based virtual machine newly implemented in
major browsers. It is well-suited for the purposes of smart contract
execution due to its sandboxed design, growing popularity, and tool
support.

Andi Gunderson wrote substantial portions of this RFC, including the
majority of the Reference-level explanation.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>