Step down as Maintainer
Due to a number of personal reasons, I am no longer able to contribute to the Hyperledger Sawtooth project and hereby step down as a maintainer. I wish the best for the future of the project and all the amazing people who have put so much work into building, supporting, and extending it.