Add RFC for Event Processor SDK
This RFC proposes extensions to the Sawtooth SDK's to support event
processing.  These events are emitted by the validator on block commits
and can be subscribed to via client messages.

These SDK extensions propose a simplified abstraction on top of these
client messages in order to provide a simple model for handling the
events.  This API is akin to the Transaction Handler/Transaction
Processor API's.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>